requirements.txt
livekit-agents
livekit-plugins-google
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
